# Code/Frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProgressResponse": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "integer"
          },
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "finished": {
            "title": "Finished",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "percent_complete": {
            "title": "Percent Complete",
            "type": "number"
          },
          "remaining": {
            "title": "Remaining",
            "type": "integer"
          },
          "running": {
            "title": "Running",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "completed",
          "failed",
          "skipped",
          "cancelled",
          "running",
          "pending",
          "finished",
          "remaining",
          "percent_complete"
        ],
        "title": "ProgressResponse",
        "type": "object"
      },
...
      "RunItemResponse": {
        "properties": {
          "call_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Id"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "reference",
          "title",
          "status",
          "call_id",
          "message",
          "started_at",
          "finished_at",
          "duration_ms"
        ],
        "title": "RunItemResponse",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "can_resume": {
            "title": "Can Resume",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "force": {
            "title": "Force",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressResponse"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "model",
          "force",
          "status",
          "message",
          "created_at",
          "started_at",
          "finished_at",
          "is_active",
          "can_resume",
          "progress",
          "items"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunSummaryResponse": {
        "properties": {
          "can_resume": {
            "title": "Can Resume",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "force": {
            "title": "Force",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressResponse"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "model",
          "force",
          "status",
          "message",
          "created_at",
          "started_at",
          "finished_at",
          "is_active",
          "can_resume",
          "progress"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
...
      "StartRunRequest": {
        "properties": {
          "acknowledge_cost": {
            "default": false,
            "description": "Required to run against a provider that charges per token.",
            "title": "Acknowledge Cost",
            "type": "boolean"
          },
          "force": {
            "default": false,
            "description": "Re-analyse calls that already have an analysis, replacing them.",
            "title": "Force",
            "type": "boolean"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model override for this run.",
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider name. Defaults to the configured provider.",
            "title": "Provider"
          }
        },
        "title": "StartRunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Call intelligence for member services.",
    "title": "NanoVox API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/corpus/runs": {
      "get": {
        "operationId": "list_runs_api_v1_corpus_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryResponse"
                  },
                  "title": "Response List Runs Api V1 Corpus Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List recent corpus runs",
        "tags": [
          "corpus"
        ]
      },
      "post": {
        "operationId": "start_run_api_v1_corpus_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a corpus run",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/v1/corpus/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_corpus_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One corpus run and every call in it",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/v1/corpus/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_corpus_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask a run to stop after the call in flight",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/v1/corpus/runs/{run_id}/resume": {
      "post": {
        "operationId": "resume_run_api_v1_corpus_runs__run_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Continue a run that was cancelled or interrupted",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/v1/corpus/runs/{run_id}/stream": {
      "get": {
        "operationId": "stream_run_api_v1_corpus_runs__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Live progress for a run, as Server-Sent Events",
        "tags": [
          "corpus"
        ]
      }
    },
...
  }
}